gnt/data/download/requirements.txt
google-cloud-storage
selenium
beautifulsoup4
requests
python-dotenv
pytest
requests-mock